requirements.txt

selenium
webdriver-manager
pandas
openpyxl
python-dotenv
